+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 8.3.0 IN PROGRESS
 
+* create Jest/RTL test for UserDetail.js. Refs UIU-2421
+
 ## [8.2.0](https://github.com/folio-org/ui-users/tree/v8.2.0) (2022-10-24)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v8.2.0)